learning_service/requirements.txt
# learning_service/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
python-jose==3.3.0
httpx==0.27.0
